Reduce clock drift; display noise
millis() was drifting because of very frequent writes to
neopixels (because the library turnsa off interrupts when
show() is called).

Also, very frequent updates to the SevenSegmentDisplay was
making it glitch (because of continuous latch switching).

Added checks to update SevenSegmentDisplay only when
necessary, and added 100ms refresh for neopixels.