Change Modes to statically allocated objects
The pervious way of doing modes was because they were
derived from a project with not enough memory to hold
all required modes, so we used dynamic allocation and
pointers.  However, that's not the typical Arduino way
of doing things, so I've change modes to have a single
instance as per more typical practise.

This will also help accessing the class from other classes
Implicit singletons kindof.