fix DEBUG #define hang
The Debugging output in the WaitMode constructor was causing a hang
because it writes to serial before Serial.begin() is called. This
debug was left over from the old way of creating the Mode objects
(where the object wasn't constructed until setup(), as opposed to
the more Arduino-typical way of having a global staticlly allocated
object).